Extract buildResponseHeaders Method
As noted on the open header issue raised by another user, parsing headers is broken for headers that may have multiple values. For example, a Symfony or Laravel application that sets two cookies, thus creating two `Set-Cookie` headers in the response. The last `Set-Cookie` header will be the only one included in the response.

While this does not fix that bug, which would be a larger breaking change since each header key would need to be an array of values, it does extract the header parsing logic into a separate method so that users may customize and thus fix the logic in their own applications if needed until a fix is provided in a major version.

Currently, I have to fix this behavior by overriding the entire `send` method.